Fixed bug 564. The problem is that 'varnum' was being cast from unsigned long to int without warning. Some debug output added, needs to be removed.
git-svn-id: https://svn.ascend4.org/trunk@2878 f15cd148-d149-4907-926e-73cf92b3ee03

+++ ascend/compiler/relation_util.h
@@ -684,7 +684,7 @@ ASC_DLLSPEC enum safe_err RelationCalcHessianMtxSafe(struct Instance *i, hessian
 double *RelationFindRoots(struct Instance *i,
         double lower_bound, double upper_bound,
         double nominal, double tolerance,
-        unsigned long *varnum,
+        int *varnum,
         int *able,
         int *nsolns);
 /**<
